Implement RotatingServerAdvice enhancements
