+++ spring-integration-ftp/src/test/java/org/springframework/integration/ftp/inbound/RotatingServersTests.java
@@ -43,8 +43,8 @@ import org.springframework.integration.dsl.IntegrationFlows;
 import org.springframework.integration.dsl.MessageChannels;
 import org.springframework.integration.dsl.Pollers;
 import org.springframework.integration.dsl.StandardIntegrationFlow;
+import org.springframework.integration.file.remote.aop.KeyDirectory;
 import org.springframework.integration.file.remote.aop.RotatingServerAdvice;
-import org.springframework.integration.file.remote.aop.RotatingServerAdvice.KeyDirectory;
 import org.springframework.integration.file.remote.session.CachingSessionFactory;
 import org.springframework.integration.file.remote.session.DefaultSessionFactoryLocator;
 import org.springframework.integration.file.remote.session.DelegatingSessionFactory;
